+++ src/locale/NodeTexts.ts
@@ -170,7 +170,7 @@ type NodeTexts = {
             DuplicateName: {
                 conflict: ConflictText;
                 resolution: Template;
-            }
+            };
             /** When a shared bind has a duplicate name that's shared. Description inputs: $1: The duplicate */
             DuplicateShare: ConflictText;
             /**
@@ -636,6 +636,7 @@ type NodeTexts = {
              */
             UnimplementedInterface: InternalConflictText;
         }>;
+    StructureDefinitionType: DescriptiveNodeText;
     /**
      * A table literal, e.g., `⎡a•# b•#⎦⎡1 2⎦`
      * Description inputs: $1 = the number of rows