+++ src/nodes/PropertyBind.ts
@@ -28,9 +28,9 @@ import { buildBindings } from './Evaluate';
 import ExceptionValue from '@values/ExceptionValue';
 import Evaluation from '@runtime/Evaluation';
 import StartEvaluation from '@runtime/StartEvaluation';
-import StructureType from './StructureType';
 import Bind from './Bind';
 import InvalidProperty from '@conflicts/InvalidProperty';
+import StructureDefinitionType from './StructureDefinitionType';
 
 export default class PropertyBind extends Expression {
     readonly reference: PropertyReference;
@@ -116,12 +116,14 @@ export default class PropertyBind extends Expression {
         const bind = this.reference.resolve(context);
         if (
             bind instanceof Bind &&
-            structureType instanceof StructureType &&
-            !structureType.definition.inputs.some((input) =>
+            structureType instanceof StructureDefinitionType &&
+            !structureType.type.definition.inputs.some((input) =>
                 input.names.sharesName(bind.names),
             )
         )
-            conflicts.push(new InvalidProperty(structureType.definition, this));
+            conflicts.push(
+                new InvalidProperty(structureType.type.definition, this),
+            );
 
         return conflicts;
     }